test: running integration tests in memory

+++ tests/vpx_read_extract_assemble_write_compare_all.rs
@@ -5,12 +5,14 @@ mod common;
 mod test {
 
     use pretty_assertions::assert_eq;
+    // TODO once we can capture logs per extract / assemble we can re-enable parallel tests
     // use rayon::prelude::*;
     use crate::common::{assert_equal_vpx, find_files, tables_dir};
     use log::info;
     use std::io;
     use std::path::{Path, PathBuf};
     use testdir::testdir;
+    use vpin::filesystem::{FileSystem, MemoryFileSystem, RealFileSystem};
 
     fn init() {
         let _ = env_logger::builder()
@@ -63,30 +65,41 @@ mod test {
             // panic!("stop");
             // if all is good we remove the test file and the extracted dir
             std::fs::remove_file(&test_vpx)?;
-            std::fs::remove_dir_all(extracted)?;
+            if let Some(extracted) = extracted {
+                std::fs::remove_dir_all(extracted)?;
+            }
             Ok(())
         })
     }
 
     struct ReadAndWriteResult {
-        extracted: PathBuf,
+        /// only set if extracted to real filesystem
+        extracted: Option<PathBuf>,
         test_vpx: PathBuf,
     }
 
     fn read_and_write_vpx(dir: &Path, path: &Path) -> io::Result<ReadAndWriteResult> {
         let original = vpin::vpx::read(path)?;
-        let extract_dir = dir.join("extracted");
-        // make dir
-        std::fs::create_dir_all(&extract_dir)?;
-        vpin::vpx::expanded::write(&original, &extract_dir).map_err(io::Error::other)?;
-        let expanded_read = vpin::vpx::expanded::read(&extract_dir).map_err(io::Error::other)?;
+        const IN_MEMORY: bool = true;
+        let (fs, extract_dir): (Box<dyn FileSystem>, PathBuf) = if IN_MEMORY {
+            (Box::new(MemoryFileSystem::new()), PathBuf::from("/vpx"))
+        } else {
+            let extract_dir = dir.join("extracted");
+            // make dir
+            std::fs::create_dir_all(&extract_dir)?;
+            (Box::new(RealFileSystem), extract_dir)
+        };
+
+        vpin::vpx::expanded::write_fs(&original, &extract_dir, &*fs).map_err(io::Error::other)?;
+        let expanded_read =
+            vpin::vpx::expanded::read_fs(&extract_dir, &*fs).map_err(io::Error::other)?;
         // special case for comparing code
         assert_eq!(original.gamedata.code, expanded_read.gamedata.code);
         let file_name = path.file_name().unwrap();
         let test_vpx_path = dir.join(file_name);
         vpin::vpx::write(&test_vpx_path, &expanded_read)?;
         Ok(ReadAndWriteResult {
-            extracted: extract_dir,
+            extracted: if IN_MEMORY { None } else { Some(extract_dir) },
             test_vpx: test_vpx_path,
         })
     }
